backend/pyproject.toml
[project]
name = "web-research-mcp-backend"
version = "0.1.0"
description = "FastAPI backend for Web Research MCP"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.1",
  "email-validator>=2.2.0",
  "beautifulsoup4>=4.12.3",
  "fastapi==0.135.2",
  "greenlet>=3.2.4",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.6.1",
  "psycopg[binary]==3.3.3",
  "PyJWT>=2.10.1",
  "requests>=2.32.3",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[tool.uv]
package = false
